tests/Element-hasAttributeNS.html

<!DOCTYPE html>
<title>Element.hasAttributeNS</title>
<script src="testharness.js"></script>
<div id="log"></div>
<script>
test(function() {
  var el = document.createElement("foo")
  assert_equals(el.hasAttributeNS(null, "bar"), false)

  el.setAttributeNS("ab", "attr", "t1")
  el.setAttributeNS("kl", "attr", "t2")
  assert_equals(el.hasAttributeNS("ab", "attr"), true)
  assert_equals(el.hasAttributeNS("kl", "attr"), true)

  el.setAttribute("style", "color:#fff;")
  assert_equals(el.hasAttributeNS(null, "style"), true)

  el = document.createElement("div")
  el.setAttributeNS("", "ALIGN", "left")
  assert_equals(el.hasAttributeNS(null, "ALIGN"), true)
  assert_equals(el.getAttribute(null, "align"), false)

  el.setAttributeNS("", "CHEEseCaKe", "tasty")
  assert_equals(el.hasAttributeNS("", "CHEESECAKE"), false)
  assert_equals(el.hasAttributeNS("", "CHEEseCaKe"), true)
  assert_equals(el.hasAttributeNS(null, "cheesecake"), false)

  el = document.createElement("div")
  el.setAttributeNS(null, "align", "left")
  el.setAttributeNS("xx", "align", "right")
  el.setAttributeNS(null, "foo", "left")
  el.setAttributeNS("xx", "foo", "right")
  document.body.appendChild(el)
  assert_equals(el.hasAttributeNS("xx", "align"), true)
  assert_equals(el.hasAttributeNS(null, "foo"), true)
})
</script>
